+++ src/main/java/server/History.java
@@ -16,8 +16,8 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="amount" type="{http://www.w3.org/2001/XMLSchema}int"/>
- *         &lt;element name="balanceAfterOperation" type="{http://www.w3.org/2001/XMLSchema}int"/>
+ *         &lt;element name="amount" type="{http://www.w3.org/2001/XMLSchema}long" minOccurs="0"/>
+ *         &lt;element name="balanceAfterOperation" type="{http://www.w3.org/2001/XMLSchema}long" minOccurs="0"/>
  *         &lt;element name="source" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
  *         &lt;element name="title" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
  *       &lt;/sequence>
@@ -37,40 +37,56 @@ import javax.xml.bind.annotation.XmlType;
 })
 public class History {
 
-    protected int amount;
-    protected int balanceAfterOperation;
+    protected Long amount;
+    protected Long balanceAfterOperation;
     protected String source;
     protected String title;
 
     /**
      * Gets the value of the amount property.
      * 
+     * @return
+     *     possible object is
+     *     {@link Long }
+     *     
      */
-    public int getAmount() {
+    public Long getAmount() {
         return amount;
     }
 
     /**
      * Sets the value of the amount property.
      * 
+     * @param value
+     *     allowed object is
+     *     {@link Long }
+     *     
      */
-    public void setAmount(int value) {
+    public void setAmount(Long value) {
         this.amount = value;
     }
 
     /**
      * Gets the value of the balanceAfterOperation property.
      * 
+     * @return
+     *     possible object is
+     *     {@link Long }
+     *     
      */
-    public int getBalanceAfterOperation() {
+    public Long getBalanceAfterOperation() {
         return balanceAfterOperation;
     }
 
     /**
      * Sets the value of the balanceAfterOperation property.
      * 
+     * @param value
+     *     allowed object is
+     *     {@link Long }
+     *     
      */
-    public void setBalanceAfterOperation(int value) {
+    public void setBalanceAfterOperation(Long value) {
         this.balanceAfterOperation = value;
     }
